+++ dcpp/NmdcHub.h
@@ -83,7 +83,7 @@ private:
 	FloodMap seekers;
 	FloodMap flooders;
 
-	uint32_t lastProtectedIPsUpdate;
+	uint64_t lastProtectedIPsUpdate;
 	StringList protectedIPs;
 
 	NmdcHub(const string& aHubURL);
@@ -120,8 +120,8 @@ private:
 	virtual string checkNick(const string& aNick);
 
 	// TimerManagerListener
-	virtual void on(Second, uint32_t aTick) throw();
-	virtual void on(Minute, uint32_t aTick) throw();
+	virtual void on(Second, uint64_t aTick) throw();
+	virtual void on(Minute, uint64_t aTick) throw();
 
 	virtual void on(Connected) throw();
 	virtual void on(Line, const string& l) throw();
